+++ lib/processMessage.js
@@ -1,5 +1,13 @@
 'use strict';
 
+const {
+  genKeyHandler,
+  xOffsetUpdate,
+  getX,
+  getT,
+  keyBindo
+} = require('../../doppler/lib');
+
 let processMessages = (message, wso) => {
   console.log('Process msg: ', message);
   try {
@@ -26,7 +34,7 @@ let setTitle = (wso, message) => {
 // eslint-disable-next-line no-unused-vars
 let goto = (wso, message) => {
   const { sidebarWidth, width } = wso.pstate;
-  const { mouse, xOffsetUpdate, getX, getT, pstate, render } = wso;
+  const { pstate, render } = wso;
 
   let currentTime = getT(sidebarWidth, pstate);
   let t = Number(message.time);
@@ -35,17 +43,12 @@ let goto = (wso, message) => {
   xOffsetUpdate(pstate, x);
   render();
   pstate.xCursor = sidebarWidth + width / 2;
-  mouse.handler({ clientX: pstate.xCursor });
 };
 
 // eslint-disable-next-line no-unused-vars
 let keyaction = (wso, message) => {
-  wso.genKeyHandler.executeKeyHandler(
-    message.key,
-    wso.keyBindo,
-    wso.pstate,
-    wso.cm
-  ) && wso.render();
+  genKeyHandler.executeKeyHandler(message.key, keyBindo, wso.pstate, wso.cm) &&
+    wso.render();
 };
 
 module.exports = { processMessages };
