+++ lib/pv2.js
@@ -18,10 +18,6 @@ const {
   // mountCodeMirror5,
   genKeyHandler,
   genOnWheel,
-  xOffsetUpdate,
-  getX,
-  getT,
-  mouse
 } = require('../../doppler/lib');
 
 const {mountCodeMirror6} = require('waveql');
@@ -84,18 +80,12 @@ const getHandler = (content, inst) => async readers => {
       render
     );
     container.elo.container.addEventListener('keydown', genKeyHandler.genKeyHandler(content, container.pstate, render, cm, keyBindo));
-    container.elo.container.addEventListener('wheel',genOnWheel(content, container.pstate, render, cm, keyBindo));
-
+    container.elo.container.addEventListener('wheel', genOnWheel(content, container.pstate, render, cm, keyBindo));
+    
     websocket.setup({
       render: render,
       pstate: container.pstate,
-      xOffsetUpdate: xOffsetUpdate,
-      getX: getX,
-      getT: getT,
-      cm: cm,
-      keyBindo: keyBindo,
-      genKeyHandler: genKeyHandler,
-      mouse: mouse
+      container: container.elo.container
     });
   });
