+++ lib/vcdrom.js
@@ -71,7 +71,6 @@ const getHandler = (content, inst) => async readers => {
     //     });
     //   });
 
-
     const cm = mountCodeMirror6(
       container.elo.sidebar,
       deso,
