+++ lib/wsClient.js
@@ -1,7 +1,8 @@
 'use strict';
 
 const processMessage = require('./processMessage.js');
-const instAtTime = require('./instructions-at-time');
+
+const { getT } = require('../../doppler/lib');
 
 let websocket;
 
@@ -56,8 +57,10 @@ function waitForSocketConnection(socket, callback) {
   }, 10); // wait a bit longer for the connection...
 }
 
-const setup = (state) => {
-  wso = state;
+const setup = (stuffWeNeed) => {
+  wso = stuffWeNeed;
+  wso.container.addEventListener('click', cursorClickHandler());
+  wso.container.addEventListener('click', instClickHandler());
 };
 
 const wsRenderPlugin = (desc, pstate /* , els */) => {
@@ -71,8 +74,7 @@ const wsRenderPlugin = (desc, pstate /* , els */) => {
   }
 };
 
-const keyPlugin = (key, event) => {
-  console.log(key);
+const keyPlugin = (key /*, event */) => {
   websocket.send(
     JSON.stringify({
       command: 'keyAction',
@@ -81,8 +83,60 @@ const keyPlugin = (key, event) => {
   );
 };
 
-const mousePlugin = (event, pstate) => {
-  console.log('mouse pluging');
-  instAtTime(event, pstate, websocket);
+const cursorClickHandler = () => (event) => {
+  const x = wso.pstate.xCursor;
+  const time = getT(x, wso.pstate);
+  console.log('cursor click at', x, time);
 };
-module.exports = { init, wsRenderPlugin, keyPlugin, mousePlugin, setup };
+
+const instClickHandler = () => (event) => {
+  var pcAddrs = document.getElementsByClassName('pc-addr');
+  const yCursor = event.clientY;
+  const x = wso.pstate.xCursor;
+  const count = pcAddrs.length;
+  var addrCursor = 'unknown';
+  const stageAttr = 'data-stage';
+  let blockAddresses = [];
+  //console.time();
+  for (let i = 0; i < count; i++) {
+    let y1 = pcAddrs[i].getBoundingClientRect().top;
+    let y2 = pcAddrs[i].getBoundingClientRect().bottom;
+    let addrBlock = pcAddrs[i].textContent.trim();
+    if (yCursor >= y1 && yCursor <= y2) {
+      addrCursor = addrBlock;
+    }
+
+    const elms = document.elementsFromPoint(x, y1 + (y2 - y1) / 2);
+    const blocks = elms.filter((elm) => {
+      // if (elm.className instanceof SVGAnimatedString) {
+      return elm.nodeName == 'rect' && elm.hasAttribute(stageAttr);
+    });
+
+    if (blocks.length != 0) {
+      let block = blocks[0];
+      //console.log(window.getComputedStyle(block).fill); //.backgroundColor;);
+      let obj = {
+        stage: block.getAttribute(stageAttr),
+        address: addrBlock,
+        color: window.getComputedStyle(block).fill
+      };
+      blockAddresses.push(obj);
+    }
+  }
+
+  let cmd = {
+    command: 'reportcursorstages',
+    address: addrCursor,
+    stages: blockAddresses
+  };
+
+  //console.timeEnd();
+  websocket.send(JSON.stringify(cmd));
+};
+
+module.exports = {
+  init,
+  setup,
+  wsRenderPlugin,
+  keyPlugin
+};
